+++ API/DeviceControl/ControlShockerObj.cs
@@ -18,7 +18,7 @@ public class ControlShockerObj
         public required bool Sound { get; set; }
         public required bool Vibrate { get; set; }
         public required bool Shock { get; set; }
-        public required uint? Duration { get; set; }
+        public required ushort? Duration { get; set; }
         public required byte? Intensity { get; set; }
     }
