+++ API/Models/Response/ShockerLimits.cs
@@ -7,5 +7,5 @@ public class ShockerLimits
     [Range(1, 100)]
     public required byte? Intensity { get; set; }
     [Range(300, 30000)]
-    public required uint? Duration { get; set; }
+    public required ushort? Duration { get; set; }
 }
